moved variable to utilities
