+++ src/utilities/variables.ts
@@ -1,0 +1,5 @@
+// global variable wrapper
+export * from './variables/chrono';
+export * from './variables/contracts';
+export * from './variables/links';
+export * from './variables/discord';
