--- src/utilities/variables/variables.ts
@@ -1,5 +1,0 @@
-// global variable wrapper
-export * from './chrono';
-export * from './contracts';
-export * from './links';
-export * from './discord';
